moved more of the fetcher over to the gosub_net crate
