--- crates/gosub_engine/src/net/fetch.rs
@@ -1,1 +1,0 @@
-pub use gosub_net::net::fetch::{fetch_response_complete, fetch_response_top, ResponseTop};
