--- crates/gosub_engine/src/net/pump.rs
@@ -1,1 +1,0 @@
-pub use gosub_net::net::pump::{spawn_pump, PumpCfg, PumpTargets};
