+++ crates/gosub_engine/src/net/shared_body.rs
@@ -1,1 +1,1 @@
-pub use gosub_net::net::shared_body::{ReaderOptions, SharedBody};
+pub use gosub_net::net::shared_body::SharedBody;
