+++ crates/gosub_engine/src/net/utils.rs
@@ -1,1 +1,1 @@
-pub use gosub_net::net::utils::{short_url, stream_to_bytes, Waiter};
+pub use gosub_net::net::utils::stream_to_bytes;
