+++ crates/gosub_net/Cargo.toml
@@ -22,6 +22,8 @@ futures-core = "0.3"
 futures-util = "0.3"
 parking_lot = "0.12"
 tempfile = "3.21.0"
+dashmap = "7.0.0-rc2"
+chrono = "0.4"
 
 [dev-dependencies]
 cow-utils = "0.1.3"
