+++ crates/gosub_net/src/net/mod.rs
@@ -1,5 +1,8 @@
+pub mod decision_hub;
 pub mod events;
 pub mod fetch;
+pub mod fetcher;
+pub mod fetcher_context;
 pub mod fs_utils;
 pub mod null_emitter;
 pub mod observer;
